+++ src/main/java/plus/misterplus/cindertally/helper/LifespanHelper.java
@@ -7,15 +7,15 @@ import plus.misterplus.cindertally.common.item.ItemLifeSpan;
 import plus.misterplus.cindertally.registry.CinderTallyRegistry;
 
 public class LifespanHelper {
-    public static Inventory getCinderTallyInventory(int lifespan) {
-        int decades = lifespan / ItemLifeSpan.VALUE_DECADE;
-        int years = lifespan % ItemLifeSpan.VALUE_DECADE / ItemLifeSpan.VALUE_YEAR;
-        int seasons = lifespan % ItemLifeSpan.VALUE_YEAR / ItemLifeSpan.VALUE_SEASON;
-        int months = lifespan % ItemLifeSpan.VALUE_SEASON / ItemLifeSpan.VALUE_MONTH;
-        int weeks = lifespan % ItemLifeSpan.VALUE_MONTH / ItemLifeSpan.VALUE_WEEK;
-        int days = lifespan % ItemLifeSpan.VALUE_WEEK / ItemLifeSpan.VALUE_DAY;
-        int hours = lifespan % ItemLifeSpan.VALUE_DAY / ItemLifeSpan.VALUE_HOUR;
-        int quarters = lifespan % ItemLifeSpan.VALUE_HOUR / ItemLifeSpan.VALUE_QUARTER;
+    public static Inventory getCinderTallyInventory(long lifespan) {
+        int decades = (int) (lifespan / ItemLifeSpan.VALUE_DECADE);
+        int years = (int) (lifespan % ItemLifeSpan.VALUE_DECADE / ItemLifeSpan.VALUE_YEAR);
+        int seasons = (int) (lifespan % ItemLifeSpan.VALUE_YEAR / ItemLifeSpan.VALUE_SEASON);
+        int months = (int) (lifespan % ItemLifeSpan.VALUE_SEASON / ItemLifeSpan.VALUE_MONTH);
+        int weeks = (int) (lifespan % ItemLifeSpan.VALUE_MONTH / ItemLifeSpan.VALUE_WEEK);
+        int days = (int) (lifespan % ItemLifeSpan.VALUE_WEEK / ItemLifeSpan.VALUE_DAY);
+        int hours = (int) (lifespan % ItemLifeSpan.VALUE_DAY / ItemLifeSpan.VALUE_HOUR);
+        int quarters = (int) (lifespan % ItemLifeSpan.VALUE_HOUR / ItemLifeSpan.VALUE_QUARTER);
         Inventory inventory = new Inventory(8);
         inventory.setItem(0, new ItemStack(CinderTallyRegistry.LIFESPAN_DECADE, decades));
         inventory.setItem(1, new ItemStack(CinderTallyRegistry.LIFESPAN_YEAR, years));
